chore(editor-state): address pr comments
