+++ packages/editor/packages/editor-state/src/effects/compiler.ts
@@ -79,7 +79,6 @@ export default async function compiler(store: StateManager<State>, events: Event
 
 			log(state, 'Compilation succeeded in ' + state.compiler.compilationTime.toFixed(2) + 'ms', 'Compiler');
 		} catch (error) {
-			console.log(error);
 			store.set('compiler.isCompiling', false);
 			const errorObject = error as Error & {
 				line?: { lineNumber: number };
