+++ packages/editor/packages/editor-state/src/impureHelpers/config/applyConfigToState.ts
@@ -2,17 +2,7 @@ import { StateManager } from '@8f4e/state-manager';
 
 import isPlainObject from '../../pureHelpers/isPlainObject';
 
-import type { State, Runtimes } from '../../types';
-
-/**
- * Interface for the expected config object structure
- */
-export interface ConfigObject {
-	memorySizeBytes?: number;
-	selectedRuntime?: number;
-	runtimeSettings?: Runtimes[];
-	disableAutoCompilation?: boolean;
-}
+import type { State, Runtimes, ConfigObject } from '../../types';
 
 /**
  * Applies the compiled config object to the editor state.
